+++ lang/en/passwords.php
@@ -1,0 +1,3 @@
+<?php
+
+return require dirname(__DIR__, 2).'/vendor/laravel/framework/src/Illuminate/Translation/lang/en/passwords.php';